tests/wpt/web-platform-tests/html/semantics/embedded-content/media-elements/playing-the-media-resource/playbackRate.html

<!doctype html>
<title>playbackRate</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
  var v = document.createElement('video');
  assert_equals(v.playbackRate, 1);
}, 'playbackRate initial value');

async_test(function(t) {
  var v = document.createElement('video');
  v.playbackRate = 2;
  v.addEventListener('ratechange', t.step_func(function() {
    t.done();
  }));
}, 'setting playbackRate');
</script>
